requirements.txt
openai>=1.0.0
pydantic>=2.7
pydantic-settings>=2.2
httpx>=0.27
PyYAML>=6.0
arxiv>=2.4.0
feedparser>=6.0.11
Jinja2>=3.1
tenacity>=8.2
PyMuPDF>=1.24
